requirements.txt
streamlit>=1.28.0
python-dotenv
openai
scikit-learn
pypdf
pandas
numpy
reportlab
python-docx
pytz
pyotp
